fixed auth usage in requests
